requirements.txt
absl-py==2.3.1
appnope==0.1.4
asttokens==2.4.1
bson==0.5.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
dnspython==2.7.0
docutils==0.21.2
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.1.0
fonttools==4.55.1
grpcio==1.75.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.5.0
kiwisolver==1.4.7
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplot==0.1.9
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
nest-asyncio==1.6.0
newsapi-python==0.2.7
nh3==0.2.19
nltk==3.9.1
numdifftools==0.9.41
numpy==2.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.12.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==6.32.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pygame==2.6.1
Pygments==2.18.0
pyloco==0.0.139
pymongo==4.11.3
PyMySQL==1.1.1
pyparsing==3.2.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.2.0
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
robotframework==7.2.2
robotframework-assertion-engine==3.0.3
robotframework-databaselibrary==2.1.1
robotframework-excellib==2.0.1
robotframework-pythonlibcore==4.4.1
scipy==1.15.2
setuptools==80.9.0
SimpleWebSocketServer==0.1.2
six==1.16.0
stack-data==0.6.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
textblob==0.18.0.post0
torch-tb-profiler==0.4.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tweepy==4.14.0
twine==6.0.1
typing==3.7.4.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.2.3
ushlex==0.99.1
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
